backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "OllieMH",email = "olivermh99@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "asyncpg",
    "python-multipart",
    "pydantic-ai == 0.0.19",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "chromadb",
    "beautifulsoup4",
    "httpx",
    "langchain-text-splitters",
    "google-generativeai (>=0.3.1,<1.0.0)",
    "greenlet",
    "logfire (>=4.16.0,<5.0.0)",
]

[tool.poetry]
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]
